Removed more dead JS code (MLDB-1910)
