+++ plugins/lang/js/function_js.cc
@@ -7,7 +7,6 @@
 
 #include "function_js.h"
 #include "mldb/core/function.h"
-#include "mldb/types/js/id_js.h"
 
 
 using namespace std;
