+++ plugins/lang/js/js_value.h
@@ -1,10 +1,10 @@
-// This file is part of MLDB. Copyright 2015 Datacratic. All rights reserved.
-
 /* js_value.h                                                      -*- C++ -*-
    Jeremy Barnes, 21 July 2010
    Copyright (c) 2010 Datacratic.  All rights reserved.
+   This file is part of MLDB. Copyright 2015 Datacratic. All rights reserved.
 
-   Definition of what a javascript value is.
+   Definition of what a javascript value is, in a way that we can forward
+   declare without needing to include all of v8.
 */
 
 #pragma once
