--- soa/js/testing/js_testing.mk
@@ -1,3 +1,0 @@
-# This file is part of MLDB. Copyright 2015 Datacratic. All rights reserved.
-
-#$(eval $(call test,js_call_test,arch js,boost))
