+++ soa/soa.mk
@@ -2,6 +2,5 @@
 
 # SOA makefile
 
-$(eval $(call include_sub_make,js))
 $(eval $(call include_sub_make,credentials))
 $(eval $(call include_sub_make,service))
